# operator.metadata.backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "SSV Operator API",
        "contact": {
            "name": "FileDrive",
            "url": "https://filedrive.io/",
            "email": "filedriveteam@outlook.com"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/operator/getoperator": {
            "get": {
                "description": "get getoperator",
                "produces": [
                    "application/json"
                ],
                "summary": "getoperator",
                "parameters": [
                    {
                        "type": "string",
                        "default": "true",
                        "description": "Whether the interface call is Swagger UI",
                        "name": "is_swag",
                        "in": "header"
                    },
                    {
                        "enum": [
                            "mainnet",
                            "testnet"
                        ],
                        "type": "string",
                        "default": "testnet",
                        "description": "net type",
                        "name": "NetType",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "604",
                        "description": "Operator Id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/CommonOkResp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/OperatorInfo"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/CommonErrorResp"
                        }
                    }
                }
            }
        },
        "/operator/getoperatorbykeyword": {
            "get": {
                "description": "get operator by keyword",
                "produces": [
                    "application/json"
                ],
                "summary": "GetOperatorByKeyword",
                "parameters": [
                    {
                        "type": "string",
                        "default": "true",
                        "description": "Whether the interface call is Swagger UI",
                        "name": "is_swag",
                        "in": "header"
                    },
                    {
                        "enum": [
                            "mainnet",
                            "testnet"
                        ],
                        "type": "string",
                        "default": "testnet",
                        "description": "net type",
                        "name": "NetType",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "t",
                        "description": "keyword",
                        "name": "keyword",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/CommonOkResp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/ListQueryOperatorRes"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/CommonErrorResp"
                        }
                    }
                }
            }
        },
        "/operator/gettimestamp": {
            "get": {
                "description": "get timestamp",
                "produces": [
                    "application/json"
                ],
                "summary": "timestamp",
                "parameters": [
                    {
                        "type": "string",
                        "default": "true",
                        "description": "Whether the interface call is Swagger UI",
                        "name": "is_swag",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/CommonOkResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/CommonErrorResp"
                        }
                    }
                }
            }
        },
        "/operator/list": {
            "get": {
                "description": "listOperator",
                "produces": [
                    "application/json"
                ],
                "summary": "listOperator",
                "parameters": [
                    {
                        "enum": [
                            "mainnet",
                            "testnet"
                        ],
                        "type": "string",
                        "default": "testnet",
                        "description": "net type",
                        "name": "NetType",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "page",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "size",
                        "name": "size",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/CommonOkResp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/ListOperatorRes"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/CommonErrorResp"
                        }
                    }
                }
            }
        },
        "/operator/update": {
            "post": {
                "description": "UpdateOperator",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "UpdateOperator",
                "parameters": [
                    {
                        "enum": [
                            "mainnet",
                            "testnet"
                        ],
                        "type": "string",
                        "default": "testnet",
                        "description": "net type",
                        "name": "NetType",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "true",
                        "description": "Whether the interface call is Swagger UI",
                        "name": "is_swag",
                        "in": "header"
                    },
                    {
                        "description": "Update operator information",
                        "name": "OperatorParams",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/OperatorParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "update operator response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/CommonOkResp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/OperatorInfo"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/CommonErrorResp"
                        }
                    }
                }
            }
        },
        "/operator/uploadimage": {
            "post": {
                "description": "upload image",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "upload",
                "parameters": [
                    {
                        "type": "file",
                        "description": "The maximum number of uploaded files cannot be greater than 5 MiB",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/CommonOkResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/CommonErrorResp"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "CommonErrorResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
        "CommonOkResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
        "ListOperatorRes": {
            "type": "object",
            "properties": {
                "cid": {
                    "type": "string"
                },
                "infos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/OperatorInfo"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "ListQueryOperatorRes": {
            "type": "object",
            "properties": {
                "infos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/OperatorInfo"
                    }
                }
            }
        },
        "OperatorInfo": {
            "type": "object",
            "properties": {
                "cid": {
                    "type": "string"
                },
                "cloud_provider": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "discord_url": {
                    "type": "string"
                },
                "eth1_node_client": {
                    "type": "string"
                },
                "eth2_node_client": {
                    "type": "string"
                },
                "linkedin_url": {
                    "type": "string"
                },
                "location": {
                    "description": "extern info",
                    "type": "string"
                },
                "logo": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "operator_id": {
                    "description": "base info",
                    "type": "integer"
                },
                "owner_address": {
                    "type": "string"
                },
                "signature": {
                    "type": "string"
                },
                "telegram_url": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer"
                },
                "twitter_url": {
                    "type": "string"
                },
                "website_url": {
                    "type": "string"
                }
            }
        },
        "OperatorParams": {
            "type": "object",
            "required": [
                "data",
                "signature"
            ],
            "properties": {
                "data": {
                    "type": "string",
                    "default": "{\"domain\":{\"name\":\"SSV\",\"chainId\":\"0x5\"},\"message\":{\"data\":\"{\\\"operator_id\\\":604,\\\"owner_address\\\":\\\"0xCd14f1aE89De9b9B7bAe9d7988d0Ad26F5DCCb11\\\",\\\"location\\\":\\\"\\\",\\\"cloud_provider\\\":\\\"\\\",\\\"eth1_node_client\\\":\\\"\\\",\\\"eth2_node_client\\\":\\\"\\\",\\\"description\\\":\\\"\\\",\\\"website_url\\\":\\\"\\\",\\\"twitter_url\\\":\\\"\\\",\\\"linkedin_url\\\":\\\"\\\",\\\"logo\\\":\\\"QmbFMke1KXqnYyBBWxB74N4c5SBnJMVAiMNRcGu6x1AwQH\\\",\\\"timestamp\\\":1665395268129}\"},\"primaryType\":\"Sign\",\"types\":{\"EIP712Domain\":[{\"name\":\"name\",\"type\":\"string\"},{\"name\":\"chainId\",\"type\":\"uint256\"}],\"Sign\":[{\"name\":\"data\",\"type\":\"string\"}]}}"
                },
                "signature": {
                    "type": "string",
                    "default": "4b10fd1f4ae8238befa681eb56bed774120a67405b9854670275919a4a27a0573a32711426770bc8d25623f08974eca3cdd6b1d8561f1f23d4ff5a882a08015b1b"
                }
            }
        }
    }
}